Move AIClient to client namespace
