+++ sdk/ai/azure-ai-resources/azure/ai/resources/client/__init__.py
@@ -4,12 +4,9 @@
 
 __path__ = __import__("pkgutil").extend_path(__path__, __name__)  # type: ignore
 
-import logging
+from ._ai_client import AIClient
 
-from ._telemetry import initialize_logger_info
+__all__ = [
+    "AIClient",
+]
 
-module_logger = logging.getLogger(__name__)
-initialize_logger_info(module_logger, terminator="\n")
-
-VERSION = "0.1.0"
-__version__ = VERSION
