+++ sdk/ai/azure-ai-resources/azure/ai/resources/client/_project_scope.py
