+++ sdk/ai/azure-ai-resources/azure/ai/resources/client/_scope_dependent_operations.py
